Refactor reaction-viewer
